split minidsp.py into focused layers for testability
Extract protocol/descriptor, state, and command-builder concerns into
separate modules: minidsp_descriptors.py, minidsp_state.py, and
minidsp_commands.py. minidsp.py now contains only the device driver and
WebSocket protocol classes.
